DialogPVRChannelsOSD elements missing from on screen display
The progress bar and recording icon are not visible in the Live TV on
screen channel display for a channel which is currently being recorded.

Suggest copying the focussed and item layouts completely (for easier
future maintenance) from MyPVRChannels and adjusting the widths
accordingly. The scrollbar position then also needs a slight adjustment.